# docs/swagger.yaml
basePath: /
definitions:
  dto.APIErrorResponse:
    properties:
      errors: {}
      message:
        example: Validation error
        type: string
      success:
        example: false
        type: boolean
    type: object
  dto.APIResponse:
    properties:
      data: {}
      message:
        example: Operation successful
        type: string
      success:
        example: true
        type: boolean
    type: object
  dto.CreateDropPointRequest:
    properties:
      address:
        example: Jl. Sukajadi No. 123, Bandung
        type: string
      is_active:
        example: true
        type: boolean
      latitude:
        example: -6.890123
        type: number
      longitude:
        example: 107.610123
        type: number
      name:
        example: Drop Point Sukajadi
        maxLength: 255
        minLength: 2
        type: string
    required:
    - address
    - name
    type: object
...
  dto.DropPointResponse:
    properties:
      address:
        example: Jl. Sukajadi No. 123, Bandung
        type: string
      created_at:
        type: string
      id:
        example: 1
        type: integer
      is_active:
        example: true
        type: boolean
      latitude:
        example: -6.890123
        type: number
      longitude:
        example: 107.610123
        type: number
      name:
        example: Drop Point Sukajadi
        type: string
      updated_at:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: admin@ecopoints.test
        type: string
      password:
        example: password
        type: string
    required:
    - email
    - password
    type: object
  dto.LoginResponse:
    properties:
      expires_in:
        example: 259200
        type: integer
      token:
        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        type: string
      token_type:
        example: Bearer
        type: string
      user:
        $ref: '#/definitions/dto.UserResponse'
    type: object
  dto.RegisterRequest:
    properties:
      email:
        example: johndoe@example.com
        maxLength: 255
        type: string
      name:
        example: John Doe
        maxLength: 255
        minLength: 2
        type: string
      password:
        example: secret123
        minLength: 8
        type: string
    required:
    - email
    - name
    - password
    type: object
...
  dto.UpdateDropPointRequest:
    properties:
      address:
        example: Jl. Sukajadi No. 125, Bandung
        type: string
      is_active:
        example: true
        type: boolean
      latitude:
        example: -6.890123
        type: number
      longitude:
        example: 107.610123
        type: number
      name:
        example: Drop Point Sukajadi Pusat
        type: string
    type: object
...
  dto.UserResponse:
    properties:
      created_at:
        type: string
      email:
        example: johndoe@example.com
        type: string
      id:
        example: 1
        type: integer
      name:
        example: John Doe
        type: string
      points_balance:
        example: 1500
        type: integer
      role:
        example: user
        type: string
      updated_at:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@ecopoints.test
    name: EcoPoints Dev Team
  description: REST API for EcoPoints application built with Go Gin and connected
    to existing Laravel MySQL database.
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
  termsOfService: http://swagger.io/terms/
  title: EcoPoints Golang REST API
  version: "1.0"
paths:
  /api/v1/auth/login:
    post:
      consumes:
      - application/json
      description: Authenticate user via email and Laravel bcrypt password, returns
        JWT token
      parameters:
      - description: Login Credentials
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Login successful
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.LoginResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Login user
      tags:
      - Auth
  /api/v1/auth/me:
    get:
      description: Retrieve profile of the currently authenticated user
      parameters:
      - description: 'Bearer token (Format: Bearer <token>)'
        in: header
        name: Authorization
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: User profile retrieved
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user profile
      tags:
      - Auth
  /api/v1/auth/register:
    post:
      consumes:
      - application/json
      description: Register a new user with default 'user' role
      parameters:
      - description: User Registration Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User successfully registered
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.UserResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "409":
          description: Email already exists
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Register new user
      tags:
      - Auth
  /api/v1/drop-points:
    get:
      description: Retrieve list of all drop point locations
      produces:
      - application/json
      responses:
        "200":
          description: Drop points retrieved successfully
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/dto.DropPointResponse'
                  type: array
              type: object
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get all drop points
      tags:
      - Drop Points
    post:
      consumes:
      - application/json
      description: Add a new drop point location (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Drop Point Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.CreateDropPointRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Drop point created
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DropPointResponse'
              type: object
        "400":
          description: Validation error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new drop point
      tags:
      - Drop Points
  /api/v1/drop-points/{id}:
    delete:
      description: Delete drop point by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Drop Point ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Drop point deleted
          schema:
            $ref: '#/definitions/dto.APIResponse'
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Drop point not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a drop point
      tags:
      - Drop Points
    get:
      description: Retrieve details of a specific drop point location
      parameters:
      - description: Drop Point ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Drop point retrieved
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DropPointResponse'
              type: object
        "400":
          description: Invalid ID
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Drop point not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      summary: Get drop point by ID
      tags:
      - Drop Points
    put:
      consumes:
      - application/json
      description: Update drop point details by ID (Admin only)
      parameters:
      - description: Bearer token
        in: header
        name: Authorization
        type: string
      - description: Drop Point ID
        in: path
        name: id
        required: true
        type: integer
      - description: Updated Drop Point Data
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateDropPointRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Drop point updated
          schema:
            allOf:
            - $ref: '#/definitions/dto.APIResponse'
            - properties:
                data:
                  $ref: '#/definitions/dto.DropPointResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "404":
          description: Drop point not found
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/dto.APIErrorResponse'
      security:
      - BearerAuth: []
      summary: Update an existing drop point
      tags:
      - Drop Points
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer " followed by your JWT token. Example: "Bearer eyJhbGciOi..."'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"